QOL changes on the backend.
Created a subfolder in content for levels along with adding to the create level method to communicate with that folder. Added comments to room and level to keep up with coding standards. Removed the unused obstacle class as it was created with the initial idea of creating a platforming game.